déplacement du code dans une classe séparée
